# dev-doc/src/data/dev-doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AddisPay Developer Platform API",
    "version": "1.0.0",
    "description": "Comprehensive REST APIs for the AddisPay Ethiopian National Payment Gateway & CMS Platform (NBE License NPS/PSO/007/2022). Supports Hosted Checkout, Direct Debit (Telebirr USSD, CBE Birr, Awash Pay, Amole), Dynamic EthQR, Android SoftPOS, Webhooks, Settlements, and Admin Content Management.",
    "contact": {
      "name": "AddisPay Developer Support",
      "email": "support@addispay.et",
      "url": "https://addispay.et"
    }
  },
  "servers": [
    {
      "url": "https://api.addispay.et/api/v1",
      "description": "Production Environment"
    },
    {
      "url": "https://sandbox.addispay.et/api/v1",
      "description": "Sandbox / Testing Environment"
    },
    {
      "url": "/api/dev-doc/simulate",
      "description": "Interactive Documentation Simulator"
    }
  ],
  "paths": {
    "/payments/checkout": {
      "post": {
        "summary": "Create Hosted Checkout Session",
        "description": "Initializes an AddisPay hosted checkout session and returns a secure payment redirect URL for the customer.",
        "tags": ["Payments & Checkout"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "currency", "orderId", "returnUrl"],
                "properties": {
                  "amount": { "type": "number", "example": 1500.00, "description": "Transaction amount in ETB or USD" },
                  "currency": { "type": "string", "enum": ["ETB", "USD"], "default": "ETB" },
                  "orderId": { "type": "string", "example": "ORD-2026-88412", "description": "Merchant unique order identifier" },
                  "description": { "type": "string", "example": "E-Commerce Purchase: Order #88412" },
                  "returnUrl": { "type": "string", "format": "uri", "example": "https://merchant.example.com/checkout/complete" },
                  "cancelUrl": { "type": "string", "format": "uri", "example": "https://merchant.example.com/checkout/cancel" },
                  "webhookUrl": { "type": "string", "format": "uri", "example": "https://merchant.example.com/api/webhooks/addispay" },
                  "customer": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string", "example": "Almaz Bekele" },
                      "email": { "type": "string", "format": "email", "example": "almaz@example.com" },
                      "phoneNumber": { "type": "string", "example": "+251911223344" }
                    }
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true,
                    "example": { "branchId": "BOLE-04", "posUser": "cashier_3" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "SUCCESS" },
                    "code": { "type": "integer", "example": 200 },
                    "data": {
                      "type": "object",
                      "properties": {
                        "sessionId": { "type": "string", "example": "chk_sess_9a8b7c6d5e4f3a2b" },
                        "checkoutUrl": { "type": "string", "example": "https://checkout.addispay.et/pay/chk_sess_9a8b7c6d5e4f3a2b" },
                        "transactionId": { "type": "string", "example": "TXN_ETB_20260827_88412" },
                        "expiresAt": { "type": "string", "format": "date-time", "example": "2026-08-27T18:25:00Z" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/payments/direct-charge": {
      "post": {
        "summary": "Direct Server-to-Server Payment",
        "description": "Debits funds directly from a mobile wallet (Telebirr USSD push, CBE Birr prompt, Awash Pay, Amole) or card token.",
        "tags": ["Payments & Checkout"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "currency", "paymentMethod", "customerPhone"],
                "properties": {
                  "amount": { "type": "number", "example": 500.00 },
                  "currency": { "type": "string", "enum": ["ETB"], "default": "ETB" },
                  "paymentMethod": { "type": "string", "enum": ["TELEBIRR_USSD", "CBE_BIRR", "AWASH_PAY", "AMOLE_DASHEN", "CARD_TOKEN"] },
                  "customerPhone": { "type": "string", "example": "+251911223344" },
                  "orderId": { "type": "string", "example": "DIR-77491" },
                  "narration": { "type": "string", "example": "Subscription Renewal Aug 2026" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Direct charge initiated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "PENDING" },
                    "code": { "type": "integer", "example": 200 },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactionId": { "type": "string", "example": "TXN_DIR_20260827_77491" },
                        "message": { "type": "string", "example": "USSD STK Push prompt sent to +251911223344. Awaiting customer PIN authorization." },
                        "status": { "type": "string", "example": "PENDING" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/payments/qr/generate": {
      "post": {
        "summary": "Generate Dynamic EthQR / Telebirr QR",
        "description": "Generates an NBE-compliant EthQR and Telebirr interoperable dynamic QR code payload with instant base64 visual output.",
        "tags": ["Payments & Checkout"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "orderId"],
                "properties": {
                  "amount": { "type": "number", "example": 450.00 },
                  "orderId": { "type": "string", "example": "QR-88219" },
                  "storeName": { "type": "string", "example": "Addis Mart Bole" },
                  "expiryMinutes": { "type": "integer", "default": 15 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "QR Code generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "SUCCESS" },
                    "code": { "type": "integer", "example": 200 },
                    "data": {
                      "type": "object",
                      "properties": {
                        "qrString": { "type": "string", "example": "00020101021226580010ET.ETHIO.ETHQR0116ADDISPAY88219020477215204581453032305406450.005802ET5914Addis Mart Bole6011Addis Ababa6304E8A2" },
                        "qrCodeDataUrl": { "type": "string", "example": "data:image/svg+xml;base64,..." },
                        "transactionId": { "type": "string", "example": "TXN_QR_20260827_88219" },
                        "expiresAt": { "type": "string", "example": "2026-08-27T18:30:00Z" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Pass your AddisPay Secret Key `sk_live_...` or JWT token in the Authorization header: `Bearer <token>`"
      },
      "SignatureAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AddisPay-Signature",
        "description": "HMAC-SHA256 timestamped signature for tamper-proof API and webhook security"
      }
    }
  }
}